+++ internal/cmd/accounts.go
@@ -42,7 +42,7 @@ var accountsListCmd = &cobra.Command{
 			return finerr.Wrap(err, finerr.CodeAuth, "token: %v", err)
 		}
 
-		client, err := plaidprov.New(c)
+		client, err := plaidprov.NewForEnv(c, c.Items[itemID].Env)
 		if err != nil {
 			return finerr.Wrap(err, finerr.CodeAuth, "client: %v", err)
 		}
@@ -80,7 +80,7 @@ var accountsGetCmd = &cobra.Command{
 		if err != nil {
 			return finerr.Wrap(err, finerr.CodeAuth, "token: %v", err)
 		}
-		client, err := plaidprov.New(c)
+		client, err := plaidprov.NewForEnv(c, c.Items[itemID].Env)
 		if err != nil {
 			return finerr.Wrap(err, finerr.CodeAuth, "client: %v", err)
 		}